requirements.txt
flask==2.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.9.0
PyMuPDF==1.23.0
pytesseract==0.3.10
Pillow==10.0.0
langchain==0.3.27
langchain-google-genai==2.1.12
langchain-community==0.3.26
chromadb==1.1.0
sentence-transformers==5.1.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
pyyaml==6.0.1
asyncio==3.4.3